requirements.txt
requests
pyperclip
PyQt5
Pillow
python-dotenv
